refactor(console): one shared bytes() formatter + deterministic e2e surface-open
Steelman pass over the memory-view polish:

- lib/format.ts bytes() absorbs OverviewPanel's pre-existing fmtBytes math
  verbatim (GB support, decimal drops at >=10 in a unit) instead of shipping a
  second near-identical formatter — exactly the drift format.ts exists to kill.
  OverviewPanel keeps only its null->em-dash guard; rendering is byte-identical.
  Unit tests cover the precision switch and GB saturation.

- e2e openMemory drops the 2s try/catch probe: the persisted uiStore hydrates
  synchronously from localStorage, so a restored surface mounts in the same
  commit that paints the rail — one instant isVisible() check decides restored
  vs. fresh (~2s faster per fresh-context test, 13 tests). The rail locator
  needs exact:true — role-name matching is substring-based and a session row
  whose topic mentions "memory" also matched once the surface was open.
